Added ability to configure MongoClientSettings
Thanks to the class JarvisFrameworkMongoClientConfigurationOptions
