+++ Jarvis.Framework.Shared/IdentitySupport/IIdentityGenerator.cs
